% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadDrugs}
\alias{loadDrugs}
\title{Load the default valid drugs dataframe}
\usage{
loadDrugs(absolute = NULL)
}
\arguments{
\item{absolute}{Optional. Absolute path to a custom \code{.rda} file containing the \code{validdrugs} dataset. If not provided, the function loads from the \code{ARTEMIS} package.}
}
\description{
Load the default valid drugs dataframe
}
