% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadGroups}
\alias{loadGroups}
\title{Load the default regimen group dataframe}
\usage{
loadGroups(absolute = NULL)
}
\arguments{
\item{absolute}{Optional. Absolute path to a custom \code{.rda} file containing the \code{validdrugs} dataset. If not provided, the function loads from the \code{ARTEMIS} package.}
}
\description{
Load the default regimen group dataframe
}
